Allow some tasks to skip semaphore + mini-refactor
